delete button event with confirm + wip notif success new contrib

+++ client/src/components/AdminEventToken.js
@@ -294,29 +294,33 @@ const AdminEventToken = param => {
 	};
 
 	const deleteEvent = async () => {
-		if (true) {
-			await fetch(
-				`http://${global.config.api.authority}/event/${param.data.id}`,
-				{
-					method: "DELETE",
-					credentials: "include"
-				}
-			)
-				.then(response => {
-					if (!response.ok) {
-						throw new Error(
-							`This is an HTTP error:
-					 The status is ${response.status}`
-						);
+		if (
+			await isConfirmed(`Desire tu supprimer l'event ${param.data.name}`)
+		) {
+			if (true) {
+				await fetch(
+					`http://${global.config.api.authority}/event/${param.data.id}`,
+					{
+						method: "DELETE",
+						credentials: "include"
 					}
-				})
-				.catch(function(error) {
-					console.log(
-						"Il y a eu un problème avec l'opération fetch: " +
-							error.message
-					);
-				});
-			window.location.reload();
+				)
+					.then(response => {
+						if (!response.ok) {
+							throw new Error(
+								`This is an HTTP error:
+					 The status is ${response.status}`
+							);
+						}
+					})
+					.catch(function(error) {
+						console.log(
+							"Il y a eu un problème avec l'opération fetch: " +
+								error.message
+						);
+					});
+				window.location.reload();
+			}
 		}
 	};
